Updated operator and solve conditionals
Included another conditional within operator to
determine if operator exists. If none, set value
of operator; if it exists, resolve num1 and num2
based on operator, then set num1 value to solution,
reset num2 value, and set operator to most recent
target.textContent.

This ensures that only two sets of numbers can be
resolved at a time. If someone presses 1 + 1 - 3,
the first pair would be resolved to 2, then it
resolves 2 - 3, and so on.